+++ CHANGELOG.md
@@ -28,6 +28,8 @@ example: * date_at - fixed some bug.
 
 ### Added
 
+* 2023-05-20 - Adicionado `User Missions` e `User Comments` na sessão de
+`Social` do admin
 * 2023-05-20 - Adicionado preview de `attachment` nos paineis do admin
 * 2023-05-19 - Adicionado novos filtros no `PostComment`
 * 2023-05-18 - Adicionado campo `thumbnail` na model MissionModel,